fix: review fixes
