+++ pragma-node/src/infra/repositories/mod.rs
@@ -2,3 +2,4 @@ pub mod entry_repository;
 pub mod onchain_repository;
 pub mod oo_repository;
 pub mod publisher_repository;
+pub mod utils;
